authentication installed

+++ main.go
@@ -2,16 +2,25 @@ package main
 
 import (
 	"encoding/json"
+	"errors"
+	"flag"
 	"fmt"
 	"html/template"
 	"log"
 	"net/http"
 	"os"
+	"strconv"
 	"strings"
+	"sync"
+	"time"
 
 	"github.com/go-chi/chi/v5"
+	"github.com/golang-jwt/jwt/v5"
+	"github.com/joho/godotenv"
+	"golang.org/x/crypto/bcrypt"
 )
 
+// missing logic in the login handler
 func main() {
 	r := chi.NewRouter()
 	apiR := chi.NewRouter()
@@ -21,15 +30,29 @@ func main() {
 		Addr:    "localhost:8080",
 		Handler: corsWrapped,
 	}
+	godotenv.Load()
 	apiCfg := &apiConfig{}
+	apiCfg.secret = os.Getenv("JWT_SECRET")
 
 	directory := http.Dir(".")
 	fsHandler := http.StripPrefix("/app", apiCfg.middlewareMetricsInc(http.FileServer(directory)))
 
+	myDB := DB{
+		path: "database.json",
+		mux:  &sync.RWMutex{},
+	}
 	//for frequent repeated tests, I make sure there IS a database, remove it, and make it again.
-	getChirps("database.json")
-	os.Remove("database.json")
-	getChirps("database.json")
+	//var dbg bool
+	//flag.BoolVar(&dbg, "debug", false, "Enable debug mode")
+
+	dbg := flag.Bool("debug", false, "Enable debug mode")
+	flag.Parse()
+
+	if *dbg {
+		myDB.LoadDB()
+		os.Remove("database.json")
+		myDB.LoadDB()
+	}
 
 	r.Mount("/api", apiR)
 	r.Mount("/admin", adminR)
@@ -37,12 +60,20 @@ func main() {
 	r.Handle("/app/*", fsHandler)
 	r.Handle("/app", fsHandler)
 	apiR.Get("/healthz", healthz)
-	//apiR.Post("/validate_chirp", validationHandler)
-	apiR.Post("/chirps", validationHandler)
-	apiR.Get("/chirps", dbRequestHandler)
+	apiR.Post("/chirps", myDB.ValidationHandler)
+	apiR.Get("/chirps", myDB.GetChirps)
+	apiR.Get("/chirps/*", myDB.GetChirp)
+	apiR.Post("/login", func(w http.ResponseWriter, r *http.Request) { LoginHandler(w, r, apiCfg, myDB) })
+	apiR.Post("/users", myDB.UserHandler)
+	apiR.Put("/users", func(w http.ResponseWriter, r *http.Request) { myDB.UpdateUser(w, r, apiCfg) })
 	adminR.Get("/metrics", func(w http.ResponseWriter, r *http.Request) { hitzHandler(w, r, apiCfg) })
 
 	httpServer.ListenAndServe()
+}
+
+type DB struct {
+	path string
+	mux  *sync.RWMutex
 }
 
 type responseRecorder struct {
@@ -52,18 +83,27 @@ type responseRecorder struct {
 
 type apiConfig struct {
 	fileserverHits int
+	secret         string
 }
 
-type chirpRAM struct {
-	Chirps map[int]chirp `json:"chirps"`
+type dbStructure struct {
+	Chirps map[int]Chirp `json:"chirps"`
+	Users  map[int]User  `json:"users"`
 }
 
-type chirp struct {
+type Chirp struct {
 	Body string `json:"body"`
 	ID   int    `json:"id"`
 }
 
+type User struct {
+	Email    string `json:"email"`
+	ID       int    `json:"id"`
+	Password string `json:"password"`
+}
+
 func healthz(w http.ResponseWriter, req *http.Request) {
+	fmt.Println("healthz hit")
 	w.Header().Add("Content-Type", "text/plain; charset=utf-8")
 	w.WriteHeader(http.StatusOK)
 	w.Write([]byte(http.StatusText(http.StatusOK)))
@@ -81,7 +121,7 @@ func hitzHandler(w http.ResponseWriter, req *http.Request, cfg *apiConfig) {
 	}{
 		Hits: cfg.fileserverHits,
 	}
-
+	fmt.Println("hitzhandler triggered")
 	w.Header().Add("Content-Type", "text/html; charset=utf-8")
 	w.WriteHeader(http.StatusOK)
 	if err := tmpl.Execute(w, data); err != nil {
@@ -120,10 +160,11 @@ func (cfg *apiConfig) middlewareMetricsInc(next http.Handler) http.Handler {
 func (r *responseRecorder) WriteHeader(status int) {
 	r.Status = status
 	r.ResponseWriter.WriteHeader(status)
+	fmt.Println("responserecorder... setup?")
 }
 
-func validationHandler(w http.ResponseWriter, r *http.Request) {
-	oldChirps, err1 := getChirps("database.json")
+func (db *DB) ValidationHandler(w http.ResponseWriter, r *http.Request) {
+	oldChirps, err1 := db.LoadDB()
 	if err1 != nil {
 		respondWithError(w, http.StatusInternalServerError, "Server experienced an error")
 		return
@@ -131,7 +172,7 @@ func validationHandler(w http.ResponseWriter, r *http.Request) {
 	chirpCount := len(oldChirps.Chirps)
 
 	decoder := json.NewDecoder(r.Body)
-	params := chirp{}
+	params := Chirp{}
 	err2 := decoder.Decode(&params)
 	if err2 != nil {
 		respondWithError(w, http.StatusInternalServerError, "Couldn't decode parameters")
@@ -148,23 +189,189 @@ func validationHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	chirpToSave := chirp{
+	chirpToSave := Chirp{
 		Body: washYourMouth(params.Body),
 		ID:   chirpCount + 1,
 	}
-	saveChirps(chirpToSave, "database.json")
+	db.NewDBEntry(chirpToSave, oldChirps)
 
 	respondWithJSON(w, 201, chirpToSave)
 }
 
-func dbRequestHandler(w http.ResponseWriter, r *http.Request) {
-	oldChirps, err := getChirps("database.json")
+func (db *DB) UserHandler(w http.ResponseWriter, r *http.Request) {
+	oldPosts, err1 := db.LoadDB()
+	if err1 != nil {
+		respondWithError(w, http.StatusInternalServerError, "Server experienced an error")
+		return
+	}
+	userCount := len(oldPosts.Users)
+
+	decoder := json.NewDecoder(r.Body)
+	params := User{}
+	err2 := decoder.Decode(&params)
+	if err2 != nil {
+		respondWithError(w, http.StatusInternalServerError, "Couldn't decode parameters")
+		return
+	}
+
+	if len(params.Email) == 0 {
+		respondWithError(w, 400, "No email detected")
+		return
+	}
+	if len(params.Password) == 0 {
+		respondWithError(w, 400, "No password detected")
+		return
+	}
+
+	hashedPassword, err3 := bcrypt.GenerateFromPassword([]byte(params.Password), 0)
+	if err3 != nil {
+		fmt.Println("failed to hash password")
+		return
+	}
+
+	userToSave := User{
+		Email:    params.Email,
+		ID:       userCount + 1,
+		Password: string(hashedPassword),
+	}
+
+	type response struct {
+		Email string `json:"email"`
+		ID    int    `json:id`
+	}
+
+	newResponse := response{
+		Email: userToSave.Email,
+		ID:    userToSave.ID,
+	}
+
+	err4 := db.NewUser(userToSave)
+	if err4 != nil {
+		respondWithError(w, 500, err4.Error())
+		return
+	}
+
+	respondWithJSON(w, 201, newResponse)
+}
+
+func LoginHandler(w http.ResponseWriter, r *http.Request, cfg *apiConfig, db DB) {
+	type LoginAttempt struct {
+		Password           string `json:"password"`
+		Email              string `json:"email"`
+		Expires_in_seconds int    `json:"expires_in_seconds"`
+	}
+
+	decoder := json.NewDecoder(r.Body)
+	params := LoginAttempt{}
+	err := decoder.Decode(&params)
+	if err != nil {
+		respondWithError(w, 500, "Can't decode request")
+		return
+	}
+
+	if len(params.Email) == 0 {
+		respondWithError(w, 400, "No email detected")
+		return
+	}
+	if len(params.Password) == 0 {
+		respondWithError(w, 400, "No password detected")
+		return
+	}
+
+	wholeDB, err := db.LoadDB()
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	allUsers := wholeDB.Users
+
+	var currentUser User
+	currentUser.ID = 0
+
+	for _, user := range allUsers {
+		if user.Email == params.Email {
+			currentUser = user
+			break
+		}
+	}
+
+	if currentUser.ID == 0 {
+		respondWithError(w, 404, "User email not found")
+		return
+	}
+	err2 := bcrypt.CompareHashAndPassword([]byte(currentUser.Password), []byte(params.Password))
+	if err2 != nil {
+		respondWithError(w, 401, "Incorrect password")
+		return
+	}
+	type response struct {
+		ID    int    `json:"id"`
+		Email string `json:"email"`
+		Token string `json:"token"`
+	}
+
+	var expiry jwt.NumericDate
+	if params.Expires_in_seconds == 0 || params.Expires_in_seconds > 86400 {
+		expiry = *jwt.NewNumericDate(time.Now().Add(time.Hour * 24))
+	} else {
+		expiry = *jwt.NewNumericDate(time.Now().Add(time.Second * time.Duration(params.Expires_in_seconds)))
+	}
+
+	newClaim := jwt.RegisteredClaims{
+		Issuer:    "chirpy",
+		IssuedAt:  jwt.NewNumericDate(time.Now()),
+		ExpiresAt: &expiry,
+		Subject:   strconv.Itoa(currentUser.ID),
+	}
+
+	newToken := jwt.NewWithClaims(jwt.SigningMethodHS256, newClaim)
+	signedToken, err3 := newToken.SignedString([]byte(cfg.secret))
+	if err3 != nil {
+		fmt.Println(err3)
+		respondWithError(w, 500, "token signing failed")
+		return
+	}
+
+	userResponse := response{
+		ID:    currentUser.ID,
+		Email: currentUser.Email,
+		Token: signedToken,
+	}
+
+	respondWithJSON(w, 200, userResponse)
+}
+
+func (db *DB) GetChirps(w http.ResponseWriter, r *http.Request) {
+	fullDB, err := db.LoadDB()
 	if err != nil {
 		respondWithError(w, http.StatusInternalServerError, "Server encountered an error")
 	}
-	payload := []chirp{}
-	for _, currentChirp := range oldChirps.Chirps {
+	payload := []Chirp{}
+	for _, currentChirp := range fullDB.Chirps {
 		payload = append(payload, currentChirp)
+	}
+
+	respondWithJSON(w, 200, payload)
+}
+
+func (db *DB) GetChirp(w http.ResponseWriter, r *http.Request) {
+	chirpIDStr := r.URL.Path[len("/api/chirps/"):]
+	chirpID, err := strconv.Atoi(chirpIDStr)
+	if err != nil {
+		respondWithError(w, http.StatusInternalServerError, "Not a valid chirp ID")
+		return
+	}
+
+	oldChirps, err := db.LoadDB()
+	if err != nil {
+		respondWithError(w, http.StatusInternalServerError, "Server encountered an error")
+	}
+
+	payload, exists := oldChirps.Chirps[chirpID]
+	if !exists {
+		respondWithError(w, 404, "Chirp not found")
+		return
 	}
 
 	respondWithJSON(w, 200, payload)
@@ -220,44 +427,178 @@ func washYourMouth(body string) string {
 	return putTogether
 }
 
-func getChirps(path string) (chirpRAM, error) {
-	//get and save should both be locking... something about a mux?
-	//I've done it before, forgotten how to do it now, and need to refresh.
-	contents, err := os.ReadFile(path)
+func (db *DB) LoadDB() (dbStructure, error) {
+	db.mux.RLock()
+	defer db.mux.RUnlock()
+	contents, err := os.ReadFile(db.path)
 	if err != nil {
 		if os.IsNotExist(err) {
 
-			saveFile, err2 := json.Marshal(chirpRAM{Chirps: map[int]chirp{}})
+			saveFile, err2 := json.Marshal(dbStructure{
+				Chirps: map[int]Chirp{},
+				Users:  map[int]User{},
+			})
 			if err2 != nil {
-				return chirpRAM{}, nil
+				return dbStructure{}, nil
 			}
-			os.WriteFile(path, saveFile, os.ModePerm)
-			return chirpRAM{}, err
+			os.WriteFile(db.path, saveFile, os.ModePerm)
+			return dbStructure{}, err
 		}
 		fmt.Println(err)
 	}
 
-	chirpsData := chirpRAM{}
+	dbData := dbStructure{}
 
-	err = json.Unmarshal([]byte(contents), &chirpsData)
+	err = json.Unmarshal([]byte(contents), &dbData)
 	if err != nil {
 		fmt.Println("Error:", err)
-		return chirpRAM{}, err
+		return dbStructure{}, err
 	}
 
-	return chirpsData, nil
+	return dbData, nil
 }
 
-func saveChirps(newChirp chirp, path string) {
-	oldChirps, err := getChirps(path)
-	if err != nil {
-		fmt.Println(err)
-	}
-	oldChirps.Chirps[newChirp.ID] = newChirp
-	saveFile, err := json.Marshal(oldChirps)
+func (db *DB) SaveDB(currentDB dbStructure) {
+	db.mux.Lock()
+	defer db.mux.Unlock()
+	saveFile, err := json.Marshal(currentDB)
 	if err != nil {
 		fmt.Println(err)
 	}
 
-	os.WriteFile(path, saveFile, os.ModePerm)
+	os.WriteFile(db.path, saveFile, os.ModePerm)
+}
+
+func (db *DB) NewDBEntry(newChirp Chirp, currentDB dbStructure) dbStructure {
+	currentDB.Chirps[newChirp.ID] = newChirp
+	db.SaveDB(currentDB)
+	return currentDB
+}
+
+func (db *DB) NewUser(newUser User) error {
+	//need to enforce email singularity
+	currentDB, err := db.LoadDB()
+	if err != nil {
+		fmt.Println(err)
+		return err
+	}
+
+	wholeDB, err2 := db.LoadDB()
+	if err2 != nil {
+		fmt.Println("failed to load database")
+	}
+	allUsers := wholeDB.Users
+	exists := false
+
+	for _, user := range allUsers {
+		if user.Email == newUser.Email {
+			exists = true
+			break
+		}
+	}
+
+	if exists {
+		err := errors.New("User already exists")
+		return err
+	}
+	currentDB.Users[newUser.ID] = newUser
+	db.SaveDB(currentDB)
+	return nil
+}
+
+func (db *DB) AuthenticateUser(w http.ResponseWriter, r *http.Request, cfg *apiConfig) (User, bool) {
+	type UpdateAccount struct {
+		Password      string `json:"password"`
+		Email         string `json:"email"`
+		Authorization string `json:"authorization"`
+	}
+
+	decoder := json.NewDecoder(r.Body)
+	params := UpdateAccount{}
+	err := decoder.Decode(&params)
+	if err != nil {
+		respondWithError(w, 500, "Can't decode request")
+		return User{}, false
+	}
+	fmt.Println(params)
+	tokenString := strings.Replace(params.Authorization, "Bearer ", "", 1)
+
+	type MyCustomClaims struct {
+		jwt.RegisteredClaims
+	}
+
+	localFunc := func(token *jwt.Token) (interface{}, error) {
+		return []byte(cfg.secret), nil
+	}
+
+	token, err2 := jwt.ParseWithClaims(tokenString, &MyCustomClaims{}, localFunc)
+	if err2 != nil {
+		fmt.Println(err2)
+		errIntro := err2.Error()[:18]
+		if errIntro == "token is malformed" {
+			fmt.Println(tokenString)
+			return User{}, false
+		}
+		if !token.Valid {
+			respondWithError(w, 401, "Unauthorized")
+		}
+		return User{}, false
+	}
+
+	id, err := token.Claims.GetSubject()
+	if err != nil {
+		respondWithError(w, 500, "Internal Server Error")
+		return User{}, false
+	}
+
+	intID, err3 := strconv.Atoi(id)
+	if err3 != nil {
+		fmt.Println("failed conversion")
+	}
+
+	authUser := User{
+		ID:       intID,
+		Email:    params.Email,
+		Password: params.Password,
+	}
+
+	return authUser, true
+}
+
+func (db *DB) UpdateUser(w http.ResponseWriter, r *http.Request, cfg *apiConfig) {
+	currentUser, ok := db.AuthenticateUser(w, r, cfg)
+	if ok {
+		wholeDB, err := db.LoadDB()
+		if err != nil {
+			fmt.Println("failed to load database")
+		}
+		allUsers := wholeDB.Users
+
+		hashedPassword, err2 := bcrypt.GenerateFromPassword([]byte(currentUser.Password), 0)
+		if err2 != nil {
+			fmt.Println("failed to hash password")
+			return
+		}
+
+		hashedUser := User{
+			ID:       currentUser.ID,
+			Email:    currentUser.Email,
+			Password: string(hashedPassword),
+		}
+
+		allUsers[currentUser.ID] = hashedUser
+		db.SaveDB(wholeDB)
+		type response struct {
+			Email    string `json:"email"`
+			Password string `json:"password"`
+		}
+
+		myResponse := response{
+			Email:    currentUser.Email,
+			Password: currentUser.Password,
+		}
+		respondWithJSON(w, 200, myResponse)
+		return
+	}
+	respondWithError(w, 401, "Token failed to resolve")
 }
